requirement.txt
numpy==1.26.4
pillow==10.4.0
PyYAML==6.0.2
torch==2.3.1+cu118
torchaudio==2.3.1+cu118
torchvision==0.18.1+cu118
tqdm==4.67.1